requirements.txt
# Polish Looted Art Discovery Engine - Dependencies

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
selenium==4.15.2

# Computer Vision & Image Processing
opencv-python>=4.12.0
Pillow>=12.0.0
imagehash>=4.3.2
scikit-image==0.22.0
scikit-learn>=1.7.2

# Machine Learning & Embeddings
transformers>=4.57.1
torch>=2.9.1
torchvision>=0.24.1
numpy>=2.2.6

# OCR
pytesseract==0.3.10

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# API Framework
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Task Queue & Caching
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
faker==20.1.0
pytest-playwright==0.4.3
playwright==1.40.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3
isort==5.13.2

# Development
ipython==8.18.1
jupyter==1.0.0
python-pptx>=1.0.0